requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
langchain-groq
langchain-community
langchain
python-multipart
pydantic[email]
supabase
gotrue
httpx
pyjwt[crypto]
sqlalchemy
psycopg2-binary
